requirements-dev.txt
Django~=4.2.7
loguru~=0.7.2
celery~=5.3.6
flower~=2.0.1
djangorestframework~=3.14.0
django-cors-headers~=4.3.1
django-mptt~=0.15.0
django-debug-toolbar~=4.2.0
djangorestframework-simplejwt~=5.3.1
drf-spectacular~=0.26.5
drf-spectacular-sidecar~=2023.10.1
django_celery_results~=2.5.1
redis~=5.0.1
requests~=2.31.0
psycopg2-binary~=2.9.9
gunicorn~=21.2.0
celery~=5.3.4
django-cors-headers~=4.3.1
django-celery-beat~=2.5.0
Pillow~=10.1.0
sqlalchemy~=2.0.23
transliterate~=1.10.2
whitenoise~=6.6.0
django-filter~=23.5
openpyxl~=3.1.2
pandas~=2.2.0
python-magic~=0.4.27
pytils~=0.4.1
phonenumbers~=8.13.30
pymorphy2~=0.9.1
django-prometheus~=2.3.1